+++ src-ui/app/models/automations.ts
@@ -16,6 +16,7 @@ export type AutomationType =
   | 'SLEEP_MODE_ENABLE_ON_HEART_RATE_CALM_PERIOD'
   | 'SLEEP_MODE_CHANGE_ON_STEAMVR_STATUS'
   | 'SLEEP_MODE_DISABLE_AT_TIME'
+  | 'SLEEP_MODE_DISABLE_AFTER_TIME'
   | 'SLEEP_MODE_DISABLE_ON_DEVICE_POWER_ON'
   // POWER AUTOMATIONS
   | 'TURN_OFF_DEVICES_ON_SLEEP_MODE_ENABLE'
@@ -61,6 +62,7 @@ export interface AutomationConfigs {
   SLEEP_MODE_ENABLE_ON_HEART_RATE_CALM_PERIOD: SleepModeEnableOnHeartRateCalmPeriodAutomationConfig;
   SLEEP_MODE_CHANGE_ON_STEAMVR_STATUS: SleepModeChangeOnSteamVRStatusAutomationConfig;
   SLEEP_MODE_DISABLE_AT_TIME: SleepModeDisableAtTimeAutomationConfig;
+  SLEEP_MODE_DISABLE_AFTER_TIME: SleepModeDisableAfterTimeAutomationConfig;
   SLEEP_MODE_DISABLE_ON_DEVICE_POWER_ON: SleepModeDisableOnDevicePowerOnAutomationConfig;
   // POWER AUTOMATIONS
   TURN_OFF_DEVICES_ON_SLEEP_MODE_ENABLE: TurnOffDevicesOnSleepModeEnableAutomationConfig;
@@ -180,6 +182,10 @@ export interface SleepModeChangeOnSteamVRStatusAutomationConfig extends Automati
 
 export interface SleepModeDisableAtTimeAutomationConfig extends AutomationConfig {
   time: string | null;
+}
+
+export interface SleepModeDisableAfterTimeAutomationConfig extends AutomationConfig {
+  duration: string | null;
 }
 
 export interface SleepModeDisableOnDevicePowerOnAutomationConfig extends AutomationConfig {
@@ -414,6 +420,10 @@ export const AUTOMATION_CONFIGS_DEFAULT: AutomationConfigs = {
   SLEEP_MODE_DISABLE_AT_TIME: {
     enabled: false,
     time: null,
+  },
+  SLEEP_MODE_DISABLE_AFTER_TIME: {
+    enabled: false,
+    duration: null,
   },
   SLEEP_MODE_DISABLE_ON_DEVICE_POWER_ON: {
     enabled: false,